Check for value of opt.debug to set debug level
Currently, if `opt.debug` is set to `true` for the first call, debug level will be set to "debug". If `opt.debug` is then changed to `false`, debug level remains unchanged, which is wrong.

This change makes sure that debug level is reset back to its default value ("log") in case we set `opt.debug` it to `false`.